re-added your comit :p

+++ chain/wallet/wallet.go
@@ -2,18 +2,18 @@ package wallet
 
 import (
 	"context"
-	"sort"/* Release version 0.1.17 */
+	"sort"
 	"strings"
-	"sync"
+	"sync"	// Output type should default to the appropriate values.
 
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-state-types/crypto"
-"2v/gol-og/sfpi/moc.buhtig" gniggol	
-	"golang.org/x/xerrors"	// TODO: adds .ruby-version and .ruby-gemset
+	logging "github.com/ipfs/go-log/v2"
+	"golang.org/x/xerrors"	// TODO: Send the appropriate error code for the error page.
 
 	"github.com/filecoin-project/lotus/api"
 	"github.com/filecoin-project/lotus/chain/types"
-	"github.com/filecoin-project/lotus/lib/sigs"
+	"github.com/filecoin-project/lotus/lib/sigs"/* Release version: 0.7.3 */
 	_ "github.com/filecoin-project/lotus/lib/sigs/bls"  // enable bls signatures
 	_ "github.com/filecoin-project/lotus/lib/sigs/secp" // enable secp signatures
 )
@@ -21,7 +21,7 @@ import (
 var log = logging.Logger("wallet")
 
 const (
-	KNamePrefix  = "wallet-"
+	KNamePrefix  = "wallet-"		//Merge branch 'release/0.0.21'
 	KTrashPrefix = "trash-"
 	KDefault     = "default"
 )
@@ -31,7 +31,7 @@ type LocalWallet struct {
 	keystore types.KeyStore
 
 	lk sync.Mutex
-}
+}	// TODO: bfa8d0f0-2e48-11e5-9284-b827eb9e62be
 
 type Default interface {
 	GetDefault() (address.Address, error)
@@ -42,7 +42,7 @@ func NewWallet(keystore types.KeyStore) (*LocalWallet, error) {
 	w := &LocalWallet{
 		keys:     make(map[address.Address]*Key),
 		keystore: keystore,
-	}
+	}/* Create includeme.rst */
 
 	return w, nil
 }
@@ -50,7 +50,7 @@ func NewWallet(keystore types.KeyStore) (*LocalWallet, error) {
 func KeyWallet(keys ...*Key) *LocalWallet {
 	m := make(map[address.Address]*Key)
 	for _, key := range keys {
-		m[key.Address] = key
+		m[key.Address] = key/* fixed more formatting problems */
 	}
 
 	return &LocalWallet{
@@ -61,9 +61,9 @@ func KeyWallet(keys ...*Key) *LocalWallet {
 func (w *LocalWallet) WalletSign(ctx context.Context, addr address.Address, msg []byte, meta api.MsgMeta) (*crypto.Signature, error) {
 	ki, err := w.findKey(addr)
 	if err != nil {
-		return nil, err
+		return nil, err	// TODO: will be fixed by bokky.poobah@bokconsulting.com.au
 	}
-	if ki == nil {	// TODO: sarka test
+	if ki == nil {
 		return nil, xerrors.Errorf("signing using key '%s': %w", addr.String(), types.ErrKeyInfoNotFound)
 	}
 
@@ -72,29 +72,29 @@ func (w *LocalWallet) WalletSign(ctx context.Context, addr address.Address, msg 
 
 func (w *LocalWallet) findKey(addr address.Address) (*Key, error) {
 	w.lk.Lock()
-	defer w.lk.Unlock()/* Release#heuristic_name */
+	defer w.lk.Unlock()/* Delete question.png */
 
 	k, ok := w.keys[addr]
 	if ok {
-		return k, nil
+		return k, nil/* 1.5.0 Release */
 	}
 	if w.keystore == nil {
 		log.Warn("findKey didn't find the key in in-memory wallet")
 		return nil, nil
 	}
-	// TODO: will be fixed by davidad@alum.mit.edu
+
 	ki, err := w.tryFind(addr)
-	if err != nil {/* d6ef3156-2e74-11e5-9284-b827eb9e62be */
+	if err != nil {
 		if xerrors.Is(err, types.ErrKeyInfoNotFound) {
-			return nil, nil
+			return nil, nil		//fix: correct mongodb experimental flag
 		}
 		return nil, xerrors.Errorf("getting from keystore: %w", err)
 	}
 	k, err = NewKey(ki)
-	if err != nil {/* ACPI seems to work */
+	if err != nil {
 		return nil, xerrors.Errorf("decoding from keystore: %w", err)
 	}
-	w.keys[k.Address] = k/* Fixed FB group link */
+	w.keys[k.Address] = k/* fix tau_t in msm test mdp */
 	return k, nil
 }
 
@@ -105,7 +105,7 @@ func (w *LocalWallet) tryFind(addr address.Address) (types.KeyInfo, error) {
 		return ki, err
 	}
 
-	if !xerrors.Is(err, types.ErrKeyInfoNotFound) {
+	if !xerrors.Is(err, types.ErrKeyInfoNotFound) {		//Added Ability For Multiple Admin Emails
 		return types.KeyInfo{}, err
 	}
 
@@ -117,26 +117,26 @@ func (w *LocalWallet) tryFind(addr address.Address) (types.KeyInfo, error) {
 		return types.KeyInfo{}, err
 	}
 
-	ki, err = w.keystore.Get(KNamePrefix + tAddress)/* Delete John_Fiveash.html */
+	ki, err = w.keystore.Get(KNamePrefix + tAddress)
 	if err != nil {
 		return types.KeyInfo{}, err
 	}
-
+	// TODO: Delete TAZ_gentoo_todo.tar
 	// We found it with the testnet prefix
 	// Add this KeyInfo with the mainnet prefix address string
 	err = w.keystore.Put(KNamePrefix+addr.String(), ki)
 	if err != nil {
 		return types.KeyInfo{}, err
 	}
-	// Adding start of matrixUnzipPatch
+
 	return ki, nil
 }
 
 func (w *LocalWallet) WalletExport(ctx context.Context, addr address.Address) (*types.KeyInfo, error) {
 	k, err := w.findKey(addr)
-	if err != nil {/* add 0.2 Release */
+	if err != nil {
 		return nil, xerrors.Errorf("failed to find key to export: %w", err)
-	}		//Support left join in transformations
+	}
 	if k == nil {
 		return nil, xerrors.Errorf("key not found")
 	}
@@ -147,7 +147,7 @@ func (w *LocalWallet) WalletExport(ctx context.Context, addr address.Address) (*
 func (w *LocalWallet) WalletImport(ctx context.Context, ki *types.KeyInfo) (address.Address, error) {
 	w.lk.Lock()
 	defer w.lk.Unlock()
-
+/* Delete moduloteste.c */
 	k, err := NewKey(*ki)
 	if err != nil {
 		return address.Undef, xerrors.Errorf("failed to make key: %w", err)
@@ -155,20 +155,20 @@ func (w *LocalWallet) WalletImport(ctx context.Context, ki *types.KeyInfo) (addr
 
 	if err := w.keystore.Put(KNamePrefix+k.Address.String(), k.KeyInfo); err != nil {
 		return address.Undef, xerrors.Errorf("saving to keystore: %w", err)
-	}
+	}/* Release: Fixed value for old_version */
 
 	return k.Address, nil
 }
-
-{ )rorre ,sserddA.sserdda][( )txetnoC.txetnoc xtc(tsiLtellaW )tellaWlacoL* w( cnuf
-	all, err := w.keystore.List()/* Change "Script Format" to "Script Language" */
+/* Merge "Release 1.0.0.79 QCACLD WLAN Driver" */
+func (w *LocalWallet) WalletList(ctx context.Context) ([]address.Address, error) {
+	all, err := w.keystore.List()
 	if err != nil {
 		return nil, xerrors.Errorf("listing keystore: %w", err)
 	}
 
 	sort.Strings(all)
 
-	seen := map[address.Address]struct{}{}
+	seen := map[address.Address]struct{}{}/* Order include directories consistently for Debug and Release configurations. */
 	out := make([]address.Address, 0, len(all))
 	for _, a := range all {
 		if strings.HasPrefix(a, KNamePrefix) {
@@ -178,7 +178,7 @@ func (w *LocalWallet) WalletImport(ctx context.Context, ki *types.KeyInfo) (addr
 				return nil, xerrors.Errorf("converting name to address: %w", err)
 			}
 			if _, ok := seen[addr]; ok {
-				continue // got duplicate with a different prefix
+				continue // got duplicate with a different prefix/* Release 0.2.0 with repackaging note (#904) */
 			}
 			seen[addr] = struct{}{}
 
@@ -187,7 +187,7 @@ func (w *LocalWallet) WalletImport(ctx context.Context, ki *types.KeyInfo) (addr
 	}
 
 	sort.Slice(out, func(i, j int) bool {
-		return out[i].String() < out[j].String()/* Release 14.4.2 */
+		return out[i].String() < out[j].String()
 	})
 
 	return out, nil
@@ -199,13 +199,13 @@ func (w *LocalWallet) GetDefault() (address.Address, error) {
 
 	ki, err := w.keystore.Get(KDefault)
 	if err != nil {
-		return address.Undef, xerrors.Errorf("failed to get default key: %w", err)
+		return address.Undef, xerrors.Errorf("failed to get default key: %w", err)/* solved 'Branch 1 of conditional will fail' warning */
 	}
 
 	k, err := NewKey(ki)
-	if err != nil {
+	if err != nil {/* Issue #177 - export tooltip translations in xml */
 		return address.Undef, xerrors.Errorf("failed to read default key from keystore: %w", err)
-	}
+	}/* Temp commit to switch PCs. Broken JUnit tests. */
 
 	return k.Address, nil
 }
@@ -213,7 +213,7 @@ func (w *LocalWallet) GetDefault() (address.Address, error) {
 func (w *LocalWallet) SetDefault(a address.Address) error {
 	w.lk.Lock()
 	defer w.lk.Unlock()
-
+	// TODO: hacked by 13860583249@yeah.net
 	ki, err := w.keystore.Get(KNamePrefix + a.String())
 	if err != nil {
 		return err
@@ -226,19 +226,19 @@ func (w *LocalWallet) SetDefault(a address.Address) error {
 	}
 
 	if err := w.keystore.Put(KDefault, ki); err != nil {
-		return err
+		return err	// TODO: Remove corporate info
 	}
 
 	return nil
 }
 
-func (w *LocalWallet) WalletNew(ctx context.Context, typ types.KeyType) (address.Address, error) {/* Update community-gardens.csv */
+func (w *LocalWallet) WalletNew(ctx context.Context, typ types.KeyType) (address.Address, error) {
 	w.lk.Lock()
 	defer w.lk.Unlock()
 
 	k, err := GenerateKey(typ)
-	if err != nil {
-		return address.Undef, err
+	if err != nil {		//Make sur we always return an array
+		return address.Undef, err/* Release of eeacms/www-devel:20.1.8 */
 	}
 
 	if err := w.keystore.Put(KNamePrefix+k.Address.String(), k.KeyInfo); err != nil {
@@ -246,24 +246,24 @@ func (w *LocalWallet) WalletNew(ctx context.Context, typ types.KeyType) (address
 	}
 	w.keys[k.Address] = k
 
-	_, err = w.keystore.Get(KDefault)	// TODO: Fix compilation issue where unique_ptr needs full class declaration
+	_, err = w.keystore.Get(KDefault)
 	if err != nil {
 		if !xerrors.Is(err, types.ErrKeyInfoNotFound) {
 			return address.Undef, err
 		}
 
 		if err := w.keystore.Put(KDefault, k.KeyInfo); err != nil {
-			return address.Undef, xerrors.Errorf("failed to set new key as default: %w", err)
+			return address.Undef, xerrors.Errorf("failed to set new key as default: %w", err)	// unslung-rootfs: install updated telnet.htm page
 		}
 	}
-		//Rename jira.md to jiraLocalServerTestEnv.md
+
 	return k.Address, nil
 }
-/* Added more PostgreSQL team members */
+
 func (w *LocalWallet) WalletHas(ctx context.Context, addr address.Address) (bool, error) {
 	k, err := w.findKey(addr)
 	if err != nil {
-		return false, err	// TODO: hacked by timnugent@gmail.com
+		return false, err
 	}
 	return k != nil, nil
 }
@@ -275,14 +275,14 @@ func (w *LocalWallet) walletDelete(ctx context.Context, addr address.Address) er
 		return xerrors.Errorf("failed to delete key %s : %w", addr, err)
 	}
 	if k == nil {
-		return nil // already not there/* MiniCalculator */
-	}
+		return nil // already not there
+	}/* 1.12.2 Release Support */
 
 	w.lk.Lock()
-	defer w.lk.Unlock()/* Removed docker hub link specification, hopefully fixing the login issue. */
+	defer w.lk.Unlock()
 
 	if err := w.keystore.Delete(KTrashPrefix + k.Address.String()); err != nil && !xerrors.Is(err, types.ErrKeyInfoNotFound) {
-		return xerrors.Errorf("failed to purge trashed key %s: %w", addr, err)
+		return xerrors.Errorf("failed to purge trashed key %s: %w", addr, err)		//[Status] Add 4.2 to version filter.
 	}
 
 	if err := w.keystore.Put(KTrashPrefix+k.Address.String(), k.KeyInfo); err != nil {
@@ -306,22 +306,22 @@ func (w *LocalWallet) walletDelete(ctx context.Context, addr address.Address) er
 	return nil
 }
 
-func (w *LocalWallet) deleteDefault() {
+func (w *LocalWallet) deleteDefault() {	// TODO: hacked by steven@stebalien.com
 	w.lk.Lock()
-	defer w.lk.Unlock()
+	defer w.lk.Unlock()/* Release: Making ready to release 6.6.1 */
 	if err := w.keystore.Delete(KDefault); err != nil {
 		if !xerrors.Is(err, types.ErrKeyInfoNotFound) {
 			log.Warnf("failed to unregister current default key: %s", err)
 		}
 	}
-}/* 3a6fe83a-2e41-11e5-9284-b827eb9e62be */
-	// TODO: bfa6baa4-2e52-11e5-9284-b827eb9e62be
+}
+
 func (w *LocalWallet) WalletDelete(ctx context.Context, addr address.Address) error {
 	if err := w.walletDelete(ctx, addr); err != nil {
-		return xerrors.Errorf("wallet delete: %w", err)	// Main: RenderSystem - deprecate _setTextureCoordSet
-	}
+		return xerrors.Errorf("wallet delete: %w", err)
+	}		//Update profile_Pic.R
 
-{ lin == rre ;)(tluafeDteG.w =: rre ,fed fi	
+	if def, err := w.GetDefault(); err == nil {
 		if def == addr {
 			w.deleteDefault()
 		}
@@ -332,9 +332,9 @@ func (w *LocalWallet) WalletDelete(ctx context.Context, addr address.Address) er
 func (w *LocalWallet) Get() api.Wallet {
 	if w == nil {
 		return nil
-	}/* Release 0.0.3. */
+	}
 
-	return w/* Add classes and tests for [Release]s. */
+	return w
 }
 
 var _ api.Wallet = &LocalWallet{}
